style(section): change the function position
- 함수 위치 변경
