+++ src/main/java/subway/service/section/SectionAddingValidation.java
@@ -25,43 +25,12 @@ public class SectionAddingValidation {
         return true;
     }
 
-    public static boolean checkExistingStationName(String stationName) {
-        List<String> stationNames = StationRepository.stationNames();
-        return stationNames.contains(stationName);
-    }
-
-    public static boolean checkExistingStationNameInLine(String lineName, String stationName) {
-        List<String> stationNameValues = new ArrayList<>();
-        LinkedList<Station> stationValues = getStationValuesInLine(lineName);
-
-        for (Station stationValue : stationValues) {
-            stationNameValues.add(stationValue.getName());
-        }
-        return stationNameValues.contains(stationName);
-    }
-
-    public static LinkedList<Station> getStationValuesInLine(String lineName) {
-        Map<Line, LinkedList<Station>> transitMaps = TransitMapRepository.transitMaps();
-        LinkedList<Station> stationValues = new LinkedList<>();
-
-        for (Map.Entry<Line, LinkedList<Station>> entry : transitMaps.entrySet()) {
-            Line key = entry.getKey();
-            String keyName = key.getName();
-            LinkedList<Station> values = entry.getValue();
-
-            if (keyName.equals(lineName)) {
-                stationValues = values;
-            }
-        }
-        return stationValues;
-    }
-
     public boolean checkStationNameValidation(String lineName, String stationName) {
-        if (!checkExistingStationName(stationName)) {
+        if (!SectionAddingService.checkExistingStationName(stationName)) {
             StationExceptionView.printInvalidStationNameExistenceException();
             return false;
         }
-        if (checkExistingStationNameInLine(lineName, stationName)) {
+        if (SectionAddingService.checkExistingStationNameInLine(lineName, stationName)) {
             SectionExceptionView.printInvalidSectionExistingStationNameInLine();
             return false;
         }
